requirements.txt
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
streamlit==0.81.1
numpy==1.20.2
matplotlib==3.4.2
pyyaml==5.4.1